+++ datadog_checks_dev/datadog_checks/dev/tooling/commands/meta/__init__.py
@@ -7,11 +7,12 @@ from ..console import CONTEXT_SETTINGS
 from .catalog import catalog
 from .changes import changes
 from .dashboard import dash
+from .jmx import jmx
 from .prometheus import prom
 from .scripts import scripts
 from .snmp import snmp
 
-ALL_COMMANDS = (catalog, changes, dash, prom, scripts, snmp)
+ALL_COMMANDS = (catalog, changes, dash, jmx, prom, scripts, snmp)
 
 
 @click.group(context_settings=CONTEXT_SETTINGS, short_help='Collection of useful utilities')