+++ datadog_checks_dev/datadog_checks/dev/tooling/config.py
@@ -123,7 +123,7 @@ def scrub_secrets(config):
             else:
                 break
 
-    for data in config['orgs'].values():
+    for data in config.get('orgs', {}).values():
         api_key = data.get('api_key')
         if api_key:
             data['api_key'] = '*' * len(api_key)